requirements.txt
asgiref==3.2.10
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.1
django-admin-sortable==2.2.3
django-environ==0.4.5
django-genericadmin==0.8.1
django-import-export==2.3.0
django-solo==1.1.3
django-tinymce4-lite==1.8.0
django-uuslug==1.2.0
django-webpack-loader==0.7.0
et-xmlfile==1.0.1
jdcal==1.4.1
jsmin==2.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==7.2.0
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
text-unidecode==1.3
xlrd==1.2.0
xlwt==1.3.0
